# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: EduSaaS - Plataforma Educacional Multi-tenant
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /auth/register:
    post:
      operationId: registerTenant
      tags: [auth]
      summary: Register new institution (tenant) with admin user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterTenantBody"
      responses:
        "201":
          description: Tenant registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
...
  /subjects:
    get:
      operationId: listSubjects
      tags: [academic]
      summary: List subjects (disciplinas)
      responses:
        "200":
          description: Subjects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Subject"
    post:
      operationId: createSubject
      tags: [academic]
      summary: Create subject
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSubjectBody"
      responses:
        "201":
          description: Subject
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subject"
  /subjects/{subjectId}:
    put:
      operationId: updateSubject
      tags: [academic]
      parameters:
        - name: subjectId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSubjectBody"
      responses:
        "200":
          description: Subject
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subject"
    delete:
      operationId: deleteSubject
      tags: [academic]
      parameters:
        - name: subjectId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /topics:
    get:
      operationId: listTopics
      tags: [academic]
      summary: List topics (matérias/conteúdos)
      parameters:
        - name: subjectId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Topics
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Topic"
    post:
      operationId: createTopic
      tags: [academic]
      summary: Create topic
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTopicBody"
      responses:
        "201":
          description: Topic
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Topic"
  /topics/{topicId}:
    put:
      operationId: updateTopic
      tags: [academic]
      parameters:
        - name: topicId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTopicBody"
      responses:
        "200":
          description: Topic
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Topic"
    delete:
      operationId: deleteTopic
      tags: [academic]
      parameters:
        - name: topicId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /series:
    get:
      operationId: listSeries
      tags: [academic]
      summary: List series (séries/anos)
      responses:
        "200":
          description: Series
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Serie"
    post:
      operationId: createSerie
      tags: [academic]
      summary: Create serie
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSerieBody"
      responses:
        "201":
          description: Serie
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Serie"
  /classes:
    get:
      operationId: listClasses
      tags: [academic]
      summary: List classes (turmas)
      parameters:
        - name: serieId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Classes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Class"
    post:
      operationId: createClass
      tags: [academic]
      summary: Create class
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateClassBody"
      responses:
        "201":
          description: Class
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Class"
  /classes/{classId}:
    get:
      operationId: getClass
      tags: [academic]
      parameters:
        - name: classId
          in: path
          required: true
          schema:
            type: integer
      summary: Get class details with students
      responses:
        "200":
          description: Class with students
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClassDetail"
    put:
      operationId: updateClass
      tags: [academic]
      parameters:
        - name: classId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateClassBody"
      responses:
        "200":
          description: Class
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Class"
    delete:
      operationId: deleteClass
      tags: [academic]
      parameters:
        - name: classId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /classes/{classId}/students:
    post:
      operationId: addStudentToClass
      tags: [academic]
      parameters:
        - name: classId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddStudentToClassBody"
      responses:
        "200":
          description: Student added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /classes/{classId}/stats:
    get:
      operationId: getClassStats
      tags: [academic]
      summary: Get class performance statistics
      parameters:
        - name: classId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Class stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClassStats"

  # ─── EXAMS ──────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success]

    # ─── AUTH ──────────────────────────────────────────────────────────────────
    LoginBody:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required: [email, password]
    RegisterTenantBody:
      type: object
      properties:
        institutionName:
          type: string
        adminName:
          type: string
        email:
          type: string
        password:
          type: string
        plan:
          type: string
          enum: [free, basic, premium]
          default: free
      required: [institutionName, adminName, email, password]
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/UserWithTenant"
      required: [token, user]

    # ─── TENANT ────────────────────────────────────────────────────────────────
    Tenant:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        logoUrl:
          type: string
          nullable: true
        primaryColor:
          type: string
          nullable: true
        plan:
          type: string
          enum: [free, basic, premium]
        educationalLevels:
          type: array
          items:
            type: string
        createdAt:
          type: string
      required: [id, name, slug, plan, educationalLevels, createdAt]
...
    User:
      type: object
      properties:
        id:
          type: integer
        tenantId:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [admin, coordinator, teacher, student]
        avatarUrl:
          type: string
          nullable: true
        registrationNumber:
          type: string
          nullable: true
        createdAt:
          type: string
      required: [id, tenantId, name, email, role, createdAt]
    UserWithTenant:
      allOf:
        - $ref: "#/components/schemas/User"
        - type: object
          properties:
            tenant:
              $ref: "#/components/schemas/Tenant"
          required: [tenant]
...
    Subject:
      type: object
      properties:
        id:
          type: integer
        tenantId:
          type: integer
        name:
          type: string
        color:
          type: string
          nullable: true
        topicsCount:
          type: integer
      required: [id, tenantId, name, topicsCount]
    CreateSubjectBody:
      type: object
      properties:
        name:
          type: string
        color:
          type: string
          nullable: true
      required: [name]
    Topic:
      type: object
      properties:
        id:
          type: integer
        subjectId:
          type: integer
        name:
          type: string
        description:
          type: string
          nullable: true
      required: [id, subjectId, name]
    CreateTopicBody:
      type: object
      properties:
        subjectId:
          type: integer
        name:
          type: string
        description:
          type: string
          nullable: true
      required: [subjectId, name]
    Serie:
      type: object
      properties:
        id:
          type: integer
        tenantId:
          type: integer
        name:
          type: string
        educationalLevel:
          type: string
          enum: [infantil, fundamental, medio, tecnico, superior]
        order:
          type: integer
      required: [id, tenantId, name, educationalLevel, order]
    CreateSerieBody:
      type: object
      properties:
        name:
          type: string
        educationalLevel:
          type: string
          enum: [infantil, fundamental, medio, tecnico, superior]
        order:
          type: integer
      required: [name, educationalLevel, order]
    Class:
      type: object
      properties:
        id:
          type: integer
        tenantId:
          type: integer
        serieId:
          type: integer
        name:
          type: string
        shift:
          type: string
          enum: [manha, tarde, noite, integral]
        year:
          type: integer
        studentsCount:
          type: integer
      required: [id, tenantId, serieId, name, shift, year, studentsCount]
    ClassDetail:
      allOf:
        - $ref: "#/components/schemas/Class"
        - type: object
          properties:
            serie:
              $ref: "#/components/schemas/Serie"
            students:
              type: array
              items:
                $ref: "#/components/schemas/User"
          required: [serie, students]
    CreateClassBody:
      type: object
      properties:
        serieId:
          type: integer
        name:
          type: string
        shift:
          type: string
          enum: [manha, tarde, noite, integral]
        year:
          type: integer
      required: [serieId, name, shift, year]
    AddStudentToClassBody:
      type: object
      properties:
        studentId:
          type: integer
      required: [studentId]
    ClassStats:
      type: object
      properties:
        classId:
          type: integer
        averageScore:
          type: number
          nullable: true
        completionRate:
          type: number
        ranking:
          type: array
          items:
            $ref: "#/components/schemas/StudentRankItem"
        bySubject:
          type: array
          items:
            $ref: "#/components/schemas/SubjectPerformance"
        examsTaken:
          type: integer
      required: [classId, completionRate, ranking, bySubject, examsTaken]

    # ─── EXAMS ─────────────────────────────────────────────────────────────────
...
    SubjectPerformance:
      type: object
      properties:
        subjectId:
          type: integer
        subjectName:
          type: string
        averageScore:
          type: number
          nullable: true
        totalAttempts:
          type: integer
        color:
          type: string
          nullable: true
      required: [subjectId, subjectName, totalAttempts]
...
    StudentRankItem:
      type: object
      properties:
        userId:
          type: integer
        name:
          type: string
        score:
          type: number
        percentage:
          type: number
        rank:
          type: integer
      required: [userId, name, score, percentage, rank]